chore(backend): Move tests & convert runtime folder to module
